refactor : 폴더명 복수 -> 단수형으로 통일
