+++ nonsoolmateClient/src/components/modal/DescriptionModal.tsx
@@ -1,4 +1,4 @@
-import Button from "@components/buttons/Button";
+import Button from "@components/button/Button";
 import { useModalDispatch } from "@hooks/useModal";
 import { DESCRIPTION_MODAL } from "constants/modal";
 import styled from "styled-components";
